Make client class lookup dynamic in factory
